created checks for pds test and decluttered oauth checks
